Added mlevel, svcol, start of density-setting

+++ universal-cslib-reflection/src/ucsl-reflection/reflections/resources/kd-tree/v1.h
@@ -17,7 +17,7 @@ namespace ucsl::resources::kd_tree::v1::reflections {
 	>;
 
 	using ResKdTreeNodeData = structure<impl::ResKdTreeNodeData, "ResKdTreeNodeData", void,
-		field<float, "deadZoneStartCoordOrLeafIndexAndNodeType">,
+		field<unsigned int, "deadZoneStartCoordOrLeafIndexAndNodeType">,
 		field<float, "deadZoneEndCoord">
 	>;
